Rename styles for consistency and add documentation
