+++ Sources/Props/Utilities/PropHelpers.swift
@@ -5,19 +5,8 @@
 //
 
 import SwiftUI
-import DynamicColor
 
-extension Color {
-    public func tinted(amount: CGFloat) -> Color {
-        Color(DynamicColor(self).tinted(amount: amount).cgColor)
-    }
-
-    public func shaded(amount: CGFloat) -> Color {
-        Color(DynamicColor(self).shaded(amount: amount).cgColor)
-    }
-}
-
-/// Duplicates the modified View once as a disabled copy of the original
+/// Duplicates the modified View once as a disabled copy of the original.
 /// This can be used to preview two copies of a styled Button, one enabled, one disabled
 public struct CloneDisabled: ViewModifier {
     public init() {}
